+++ src/sudoku_game/S_PathDifficultyAdjustor.java
@@ -35,11 +35,11 @@ class S_PathDifficultyAdjustor implements DifficultyAdjustor {
 	
 	private int determineAmountOfGivens(Random rng, int lower, int upper, int total) {
 		int percent = rng.nextInt((upper - lower) + 1) + lower;
-		return (int)Math.round(total * (percent / 100.0));
+		return Math.round(total * (percent / 100.0f));
 	}
 	
 	private int determineLowerBound(int lowerBoundOnGivensPerUnit, int dimensions) {
-		return (int)Math.round(dimensions * (lowerBoundOnGivensPerUnit / 100.0));
+		return Math.round(dimensions * (lowerBoundOnGivensPerUnit / 100.0f));
 	}
 	
 	private void performAdjustment(Board board, int amount, int lowerBound, Random rng) {
